+++ service-product/module-api/src/test/java/com/github/msafriends/serviceproduct/moduleapi/controller/internal/v1/ProductControllerTest.java
@@ -11,15 +11,18 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
+import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.ResultActions;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.msafriends.serviceproduct.moduleapi.service.product.DefaultProductServiceImpl;
 
 import com.github.msafriends.serviceproduct.common.AcceptanceTest;
 
-
+@ActiveProfiles(value = "test")
+@Transactional
 class ProductControllerTest extends AcceptanceTest {
 	@MockBean
 	DefaultProductServiceImpl defaultProductServiceImpl;
